fix(wasm): portabilidad bytecode wasm32 + parser EOF infinito + stdlib embebida
- codegen: num_cache HashMap<u64,usize> — en wasm32 usize=32 bits truncaba
  f64::to_bits() (u64) y colisionaban todos los nums enteros pequeños (targets
  de Jmp/JmpIf y argc) en key 0 → Stack underflow en matematicas_potencia
- bytecode: func.start encode como u64 — en wasm32 start.to_le_bytes() daba
  4 bytes vs 8 en x64 → .nvc no portable entre archivos (F9.1)
- parser: advance() en EOF devolvía previous_token() (el '(' de 'imprimir(')
  → parse_primary recurría infinitamente → stack overflow nativo / trap
  'memory access out of bounds' en wasm; ahora devuelve el token Eof real
  → E020 'Expresión inesperada: Eof'
- loader: ModuleLoader::with_memory_files + prefijo __lumen_mem__ — stdlib
  embebida en memoria para el playground WASM (build.rs la incrusta)
- lumen-wasm: LumenRuntime con run/check/compile_to_bytes + panic hook JS,
  Instant cfg(wasm32) en vm.rs
- repro_virtual.rs: batería stdlib con memory_files (HOLA/2/1024/E020)
- wasm-smoke-test.mjs: 12/12 verde en Node puro

+++ crates/lumen-codegen/src/codegen.rs
@@ -8,7 +8,7 @@ pub struct Codegen {
     func_starts: HashMap<String, usize>,
     string_cache: HashMap<String, usize>,
     int_cache: HashMap<i64, usize>,
-    num_cache: HashMap<usize, usize>,
+    num_cache: HashMap<u64, usize>,
     name_cache: HashMap<String, usize>,
 }
 
@@ -54,7 +54,7 @@ impl Codegen {
     }
 
     fn intern_num(&mut self, n: f64) -> usize {
-        let key = n.to_bits() as usize;
+        let key = n.to_bits();
         if let Some(&idx) = self.num_cache.get(&key) {
             idx
         } else {
